requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.8
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
dill==0.3.9
distro==1.9.0
Django==5.1.5
django-ckeditor==6.7.2
django-js-asset==3.0.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
openai==1.82.0
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pylint==3.3.4
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
selenium==4.27.1
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
tomlkit==0.13.2
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.3
watchfiles==1.0.5
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
